asdasdasda

+++ Employees/FixedSalary.cs
@@ -9,7 +9,7 @@ namespace Employees
     class FixedSalary : Person
     {
         public FixedSalary() { }
-        public FixedSalary(string name, string surname, string patronymic, int age, double salary) : base(name, surname, patronymic, age, salary) { }
+        public FixedSalary(string name, string surname, string patronymic, decimal age, double salary) : base(name, surname, patronymic, age, salary) { }
         public override double Payment()
         {
             return Salary;
